tests/wpt/tests/html/canvas/element/path-objects/2d.path.arc.nonfinite.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>Canvas test: 2d.path.arc.nonfinite</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.path.arc.nonfinite</h1>
<p class="desc">arc() with Infinity/NaN is ignored</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:<p><img src="/images/green-100x50.png" class="output expected" id="expected" alt="">
<ul id="d"></ul>
<script>
var t = async_test("arc() with Infinity/NaN is ignored");
_addTest(function(canvas, ctx) {

  ctx.fillStyle = '#f00';
  ctx.fillRect(0, 0, 100, 50);
  ctx.moveTo(0, 0);
  ctx.lineTo(100, 0);
  ctx.arc(Infinity, 0, 50, 0, 2*Math.PI, true);
  ctx.arc(-Infinity, 0, 50, 0, 2*Math.PI, true);
  ctx.arc(NaN, 0, 50, 0, 2*Math.PI, true);
  ctx.arc(0, Infinity, 50, 0, 2*Math.PI, true);
  ctx.arc(0, -Infinity, 50, 0, 2*Math.PI, true);
  ctx.arc(0, NaN, 50, 0, 2*Math.PI, true);
  ctx.arc(0, 0, Infinity, 0, 2*Math.PI, true);
  ctx.arc(0, 0, -Infinity, 0, 2*Math.PI, true);
  ctx.arc(0, 0, NaN, 0, 2*Math.PI, true);
  ctx.arc(0, 0, 50, Infinity, 2*Math.PI, true);
  ctx.arc(0, 0, 50, -Infinity, 2*Math.PI, true);
  ctx.arc(0, 0, 50, NaN, 2*Math.PI, true);
  ctx.arc(0, 0, 50, 0, Infinity, true);
  ctx.arc(0, 0, 50, 0, -Infinity, true);
  ctx.arc(0, 0, 50, 0, NaN, true);
  ctx.arc(Infinity, Infinity, 50, 0, 2*Math.PI, true);
  ctx.arc(Infinity, Infinity, Infinity, 0, 2*Math.PI, true);
  ctx.arc(Infinity, Infinity, Infinity, Infinity, 2*Math.PI, true);
  ctx.arc(Infinity, Infinity, Infinity, Infinity, Infinity, true);
  ctx.arc(Infinity, Infinity, Infinity, 0, Infinity, true);
  ctx.arc(Infinity, Infinity, 50, Infinity, 2*Math.PI, true);
  ctx.arc(Infinity, Infinity, 50, Infinity, Infinity, true);
  ctx.arc(Infinity, Infinity, 50, 0, Infinity, true);
  ctx.arc(Infinity, 0, Infinity, 0, 2*Math.PI, true);
  ctx.arc(Infinity, 0, Infinity, Infinity, 2*Math.PI, true);
  ctx.arc(Infinity, 0, Infinity, Infinity, Infinity, true);
  ctx.arc(Infinity, 0, Infinity, 0, Infinity, true);
  ctx.arc(Infinity, 0, 50, Infinity, 2*Math.PI, true);
  ctx.arc(Infinity, 0, 50, Infinity, Infinity, true);
  ctx.arc(Infinity, 0, 50, 0, Infinity, true);
  ctx.arc(0, Infinity, Infinity, 0, 2*Math.PI, true);
  ctx.arc(0, Infinity, Infinity, Infinity, 2*Math.PI, true);
  ctx.arc(0, Infinity, Infinity, Infinity, Infinity, true);
  ctx.arc(0, Infinity, Infinity, 0, Infinity, true);
  ctx.arc(0, Infinity, 50, Infinity, 2*Math.PI, true);
  ctx.arc(0, Infinity, 50, Infinity, Infinity, true);
  ctx.arc(0, Infinity, 50, 0, Infinity, true);
  ctx.arc(0, 0, Infinity, Infinity, 2*Math.PI, true);
  ctx.arc(0, 0, Infinity, Infinity, Infinity, true);
  ctx.arc(0, 0, Infinity, 0, Infinity, true);
  ctx.arc(0, 0, 50, Infinity, Infinity, true);
  ctx.lineTo(100, 50);
  ctx.lineTo(0, 50);
  ctx.fillStyle = '#0f0';
  ctx.fill();
  _assertPixel(canvas, 50,25, 0,255,0,255);
  _assertPixel(canvas, 90,45, 0,255,0,255);

});
</script>
