LayoutTests/imported/w3c/web-platform-tests/html/canvas/element/path-objects/2d.path.arc.nonfinite-expected.txt

2d.path.arc.nonfinite
arc() with Infinity/NaN is ignored
Actual output:
Expected output:


PASS arc() with Infinity/NaN is ignored 
